add Clima modules and remember Ciudades modules, add api and models components
