+++ app/src/main/AndroidManifest.xml
@@ -26,6 +26,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".ClimaActivity" />
     </application>
 
 </manifest>
