+++ app/src/main/java/com/example/climaapp/MainActivity.kt
@@ -7,10 +7,7 @@ import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import com.example.climaapp.ui.theme.ClimaAppTheme
 
 class MainActivity : ComponentActivity() {
@@ -20,7 +17,7 @@ class MainActivity : ComponentActivity() {
         setContent {
             ClimaAppTheme {
                 Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
-                    IndexView(
+                    CiudadesView(
                         modifier = Modifier.padding(innerPadding)
                     )
                 }
