+++ app/src/main/java/com/example/climaapp/data/models/model.kt
@@ -1,0 +1,42 @@
+package com.example.climaapp.data.models
+
+data class CityGeoInfo(
+    val name: String,
+    val lat: Double,
+    val lon: Double,
+    val country: String
+)
+
+data class WeatherResponse(
+    val name: String,
+    val main: Main,
+    val weather: List<Weather>,
+    val wind: Wind
+)
+
+data class Main(
+    val temp: Double,
+    val feels_like: Double,
+    val humidity: Int
+)
+
+data class Weather(
+    val main: String,
+    val description: String,
+    val icon: String
+)
+
+data class Wind(
+    val speed: Double
+)
+
+data class ForecastResponse(
+    val list: List<ForecastItem>
+)
+
+data class ForecastItem(
+    val dt: Long,
+    val main: Main,
+    val weather: List<Weather>
+)
+
